+++ scanpipe/tests/test_models.py
@@ -32,12 +32,15 @@ from unittest import skipIf
 
 from django.apps import apps
 from django.core.files.uploadedfile import SimpleUploadedFile
+from django.core.management import call_command
 from django.db import DataError
 from django.db import connection
 from django.test import TestCase
 from django.test import TransactionTestCase
 from django.test import override_settings
 from django.utils import timezone
+
+from rq.job import JobStatus
 
 from scancodeio import __version__ as scancodeio_version
 from scanpipe.models import CodebaseResource
@@ -58,7 +61,10 @@ from scanpipe.tests.pipelines.do_nothing import DoNothing
 scanpipe_app = apps.get_app_config("scanpipe")
 
 
-class BaseScanPipeModelsTest:
+class ScanPipeModelsTest(TestCase):
+    data_location = Path(__file__).parent / "data"
+    fixtures = [data_location / "asgiref-3.3.0_fixtures.json"]
+
     def setUp(self):
         self.project1 = Project.objects.create(name="Analysis")
         self.project_asgiref = Project.objects.get(name="asgiref")
@@ -69,11 +75,6 @@ class BaseScanPipeModelsTest:
             pipeline_name=pipeline,
             **kwargs,
         )
-
-
-class ScanPipeModelsTest(BaseScanPipeModelsTest, TestCase):
-    data_location = Path(__file__).parent / "data"
-    fixtures = [data_location / "asgiref-3.3.0_fixtures.json"]
 
     def test_scanpipe_project_model_extra_data(self):
         self.assertEqual({}, self.project1.extra_data)
@@ -545,7 +546,9 @@ class ScanPipeModelsTest(BaseScanPipeModelsTest, TestCase):
     def test_scanpipe_run_model_queryset_methods(self):
         now = timezone.now()
 
-        running = self.create_run(pipeline="running", task_start_date=now)
+        running = self.create_run(
+            pipeline="running", task_start_date=now, task_id=uuid.uuid4()
+        )
         not_started = self.create_run(pipeline="not_started")
         queued = self.create_run(pipeline="queued", task_id=uuid.uuid4())
         executed = self.create_run(
@@ -579,6 +582,9 @@ class ScanPipeModelsTest(BaseScanPipeModelsTest, TestCase):
         qs = self.project1.runs.failed()
         self.assertQuerysetEqual(qs, [failed])
 
+        queued_or_running_qs = self.project1.runs.queued_or_running()
+        self.assertQuerysetEqual(queued_or_running_qs, [running, queued])
+
     def test_scanpipe_run_model_status_property(self):
         now = timezone.now()
 
@@ -597,6 +603,59 @@ class ScanPipeModelsTest(BaseScanPipeModelsTest, TestCase):
         self.assertEqual(Run.Status.QUEUED, queued.status)
         self.assertEqual(Run.Status.SUCCESS, succeed.status)
         self.assertEqual(Run.Status.FAILURE, failed.status)
+
+    @override_settings(SCANCODEIO_ASYNC=True)
+    @mock.patch("scanpipe.models.Run.execute_task_async")
+    @mock.patch("scanpipe.models.Run.job_status", new_callable=mock.PropertyMock)
+    def test_scanpipe_run_model_sync_with_job_async_mode(
+        self, mock_job_status, mock_execute_task
+    ):
+        queued = self.create_run(task_id=uuid.uuid4())
+        self.assertEqual(Run.Status.QUEUED, queued.status)
+        mock_job_status.return_value = None
+        queued.sync_with_job()
+        mock_execute_task.assert_called_once()
+
+        running = self.create_run(task_id=uuid.uuid4(), task_start_date=timezone.now())
+        self.assertEqual(Run.Status.RUNNING, running.status)
+        mock_job_status.return_value = None
+        running.sync_with_job()
+        running.refresh_from_db()
+        self.assertTrue(running.task_staled)
+
+        running = self.create_run(task_id=uuid.uuid4(), task_start_date=timezone.now())
+        mock_job_status.return_value = JobStatus.STOPPED
+        running.sync_with_job()
+        running.refresh_from_db()
+        self.assertTrue(running.task_stopped)
+
+        running = self.create_run(task_id=uuid.uuid4(), task_start_date=timezone.now())
+        mock_job_status.return_value = JobStatus.FAILED
+        running.sync_with_job()
+        running.refresh_from_db()
+        self.assertTrue(running.task_failed)
+        expected = "Job was moved to the FailedJobRegistry during cleanup"
+        self.assertEqual(expected, running.task_output)
+
+        running = self.create_run(task_id=uuid.uuid4(), task_start_date=timezone.now())
+        mock_job_status.return_value = "Something else"
+        running.sync_with_job()
+        running.refresh_from_db()
+        self.assertTrue(running.task_staled)
+
+    @override_settings(SCANCODEIO_ASYNC=False)
+    @mock.patch("scanpipe.models.Run.execute_task_async")
+    def test_scanpipe_run_model_sync_with_job_sync_mode(self, mock_execute_task):
+        queued = self.create_run(task_id=uuid.uuid4())
+        self.assertEqual(Run.Status.QUEUED, queued.status)
+        queued.sync_with_job()
+        mock_execute_task.assert_called_once()
+
+        running = self.create_run(task_id=uuid.uuid4(), task_start_date=timezone.now())
+        self.assertEqual(Run.Status.RUNNING, running.status)
+        running.sync_with_job()
+        running.refresh_from_db()
+        self.assertTrue(running.task_staled)
 
     def test_scanpipe_run_model_append_to_log(self):
         run1 = self.create_run()
@@ -1062,6 +1121,55 @@ class ScanPipeModelsTest(BaseScanPipeModelsTest, TestCase):
             qs = DiscoveredPackage.objects.for_package_url(purl)
             self.assertEqual(expected_count, qs.count(), msg=purl)
 
+    def test_scanpipe_codebase_resource_model_walk_method(self):
+        fixtures = self.data_location / "asgiref-3.3.0_walk_test_fixtures.json"
+        call_command("loaddata", fixtures, **{"verbosity": 0})
+        asgiref_root = self.project_asgiref.codebaseresources.get(path="codebase")
+
+        topdown_paths = list(r.path for r in asgiref_root.walk(topdown=True))
+        expected_topdown_paths = [
+            "codebase/asgiref-3.3.0.whl",
+            "codebase/asgiref-3.3.0.whl-extract",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/compatibility.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/current_thread_executor.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/local.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/server.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/sync.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/testing.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/timeout.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/wsgi.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/LICENSE",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/METADATA",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/RECORD",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/top_level.txt",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/WHEEL",
+        ]
+        self.assertEqual(expected_topdown_paths, topdown_paths)
+
+        bottom_up_paths = list(r.path for r in asgiref_root.walk(topdown=False))
+        expected_bottom_up_paths = [
+            "codebase/asgiref-3.3.0.whl",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/compatibility.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/current_thread_executor.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/local.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/server.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/sync.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/testing.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/timeout.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref/wsgi.py",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/LICENSE",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/METADATA",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/RECORD",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/top_level.txt",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/WHEEL",
+            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info",
+            "codebase/asgiref-3.3.0.whl-extract",
+        ]
+        self.assertEqual(expected_bottom_up_paths, bottom_up_paths)
+
 
 class ScanPipeModelsTransactionTest(TransactionTestCase):
     """
@@ -1230,60 +1338,3 @@ class ScanPipeModelsTransactionTest(TransactionTestCase):
         self.assertTrue(error.details["codebase_resource_pk"])
         self.assertEqual(resource.path, error.details["codebase_resource_path"])
         self.assertIn("in save", error.traceback)
-
-
-class ScanPipeWalkTest(BaseScanPipeModelsTest, TestCase):
-    data_location = Path(__file__).parent / "data"
-    fixtures = [data_location / "asgiref-3.3.0_walk_test_fixtures.json"]
-
-    def test_scanpipe_codebase_resource_walk(self):
-        fixtures = [self.data_location / "asgiref-3.3.0_walk_test_fixtures.json"]
-        project = Project.objects.create(name="asgiref_walk_test")
-        project_asgiref = Project.objects.get(name="asgiref")
-        asgiref_root = self.project_asgiref.codebaseresources.get(path="codebase")
-
-        topdown_paths = list(r.path for r in asgiref_root.walk(topdown=True))
-        expected_topdown_paths = [
-            "codebase/asgiref-3.3.0.whl",
-            "codebase/asgiref-3.3.0.whl-extract",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/compatibility.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/current_thread_executor.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/__init__.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/local.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/server.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/sync.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/testing.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/timeout.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/wsgi.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/LICENSE",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/METADATA",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/RECORD",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/top_level.txt",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/WHEEL",
-        ]
-        self.assertEqual(expected_topdown_paths, topdown_paths)
-
-        bottom_up_paths = list(r.path for r in asgiref_root.walk(topdown=False))
-        expected_bottom_up_paths = [
-            "codebase/asgiref-3.3.0.whl",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/compatibility.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/current_thread_executor.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/__init__.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/local.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/server.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/sync.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/testing.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/timeout.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref/wsgi.py",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/LICENSE",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/METADATA",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/RECORD",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/top_level.txt",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info/WHEEL",
-            "codebase/asgiref-3.3.0.whl-extract/asgiref-3.3.0.dist-info",
-            "codebase/asgiref-3.3.0.whl-extract",
-        ]
-        self.assertEqual(expected_bottom_up_paths, bottom_up_paths)